Create Add Food Allergens icons to "Price List" widget for Restaurants
"Price List" widget is a very handfull tool for Restaurants or Bar webs but the big problem is that there is no possibility to add the food allergen/intolerance icons (lactose,gluten,soja,sesame,nuts...etc) wich in Spain and most Europe is mandatory by law. It would be very helpfull if you could solve this problem and add this option to the widget Please. It would make more easy to work with this type of clients ( bars, restaurants, coffe...).
Also, in the same widget it would be helpfull too add a second price option ( for example:  one product has 2 prices: one for the inside bar and other price at the terraze ).